+++ Core/FormMpContent.cs
@@ -218,7 +218,7 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public FormMpContent Put(string name, IData v, int proj = 0x00ff)
+        public FormMpContent Put(string name, IData v, short proj = 0x00ff)
         {
             Part(name);
             if (v == null)
@@ -234,7 +234,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public FormMpContent Put<D>(string name, D[] v, int proj = 0x00ff) where D : IData
+        public FormMpContent Put<D>(string name, D[] v, short proj = 0x00ff) where D : IData
         {
             Part(name);
             if (v == null)
